# perf: simpler & faster base32 encode