--- a/.changesets/r9sme.patch.md
+++ b/.changesets/r9sme.patch.md
@@ -1,0 +1,1 @@
+Improve `base32.encode()` performance.