@pilcrowOnPaper 2024-02-07
Ah interesting - forgot that bit shifting "loops" > 32